pyproject.toml
[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyglsl"
version = "0.0.8"
authors = [
  { name="George Watson", email="gigolo@hotmail.co.uk" },
]
description = "Transform Python to GLSL"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "attrs>=23.1.0,<26.0.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
dev = [
    "PyOpenGL>=3.1",
    "glfw>=2.0",
]

[project.urls]
Homepage = "https://github.com/takeiteasy/pyglsl"
Issues = "https://github.com/takeiteasy/pyglsl/issues"

